its broken rewrite from scratch
its been almost 4 months since i have used titanium it is a broken
piece of garbage and so not worth my time.  Nothing works at all as it
should.  Oh if we could just do this using the actual iOS tools that
would be nice…. will continue and try again tomorrow