app complete
added a new window that opens and displays more detail on the reddit
post clicked on. this new window consists of a label to hold the title,
the window title bar is the author and the window also has a scrollable
web view to display the original information.  Considering assignment
complete at this point other than a few little modifications to make
use exports instead of include